+++ src/components/Header/TopSearchNavHeader.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import HeaderContainer from './HeaderContainer';
 import BackButton from './BackButton';
-import { SearchContainer, Input } from './style/HeaderStyle';
+import { SearchContainer, Input } from './HeaderStyle';
 
 const TopSearchNavHeader = ({ value, setValue }) => {
   const handleChange = (event) => {
