+++ src/components/Header/TopUploadHeader.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import HeaderContainer from './HeaderContainer';
 import BackButton from './BackButton';
 import ButtonContainer from '../common/Button/ButtonContainer';
-import { Storage } from './style/HeaderStyle';
+import { Storage } from './HeaderStyle';
 
 const TopUploadHeader = ({ text, handleClick, isDisabled }) => {
   return (
